Refactor html preview into smaller files
